+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0
  - Remove command
  - Page margins are now set using md-style, until better solution can be found
+ - `md-bib-url` now has accessed property
 
 ## 1.3.0
  - Multiple references at once!